pyproject.toml
[tool.poetry]
name = "tracker-of-habits"
version = "0.1.0"
description = ""
authors = ["Arthur Gaetov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
celery = "^5.4.0"
redis = "^5.0.7"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
black = "^24.4.2"
django-filter = "^24.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-celery-beat = "^2.6.0"
gevent = "^24.2.1"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
coverage = {extras = ["toml"], version = "^7.5.4"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    # omit anything in a .local directory anywhere
    "*/.local/*",
    # omit everything in /usr
    "/config/*",
    "*/__init__.py"
    ]